* tests/unit/: Makefile.am, capi/GEOSCoordSeq.cpp:           new experimental test for C-API GEOSCoordSeq.         * capi/geos_c.cpp (GEOSCoordSeq_create): removed           suspicious static cast.
git-svn-id: http://svn.osgeo.org/geos/trunk@1923 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ capi/geos_c.cpp
@@ -1461,7 +1461,7 @@ CoordinateSequence *
 GEOSCoordSeq_create(unsigned int size, unsigned int dims)
 {
 	try {
-		return geomFactory->getCoordinateSequenceFactory()->create(size, static_cast<int>(dims));
+		return geomFactory->getCoordinateSequenceFactory()->create(size, dims);
 	}
 
 	catch (const std::exception &e)
